+++ bee-tangle/tests/tangle.rs
@@ -5,15 +5,15 @@ use bee_tangle::Tangle;
 use bee_test::rand::message::{metadata::rand_message_metadata, rand_message};
 
 #[tokio::test]
-async fn insert() {
-    let message = rand_message();
+async fn insert_get() {
+    let (message, metadata) = (rand_message(), rand_message_metadata());
     let message_id = message.id();
-    let metadata = rand_message_metadata();
 
     let tangle = Tangle::new();
     tangle.insert(message_id, message.clone(), metadata.clone()).await;
 
     let message_data = tangle.get(&message_id).await.unwrap();
+
     assert_eq!(*message_data.message(), message);
     assert_eq!(*message_data.metadata(), metadata);
 }
